move verify_image_exif_write to geotag_from_exif
